refactor: Unify remote dataset and records class into a single RemoteFeedbackDataset and RemoteFeedbackRecords
